refactor(proptest): split op_specs into a package + comparator polish
