+++ tests/proptest/op_specs/reductions.py
@@ -1,0 +1,429 @@
+"""Spec builders for the reductions op group."""
+
+import typing as T
+
+import torch
+from hypothesis import strategies as st
+
+from torch_to_nnef.inference_target.tract import TractCheckTolerance
+
+from ...wrapper import (
+    TensorFnPrimitive,
+)
+from ..inputs import Interval, tensor_st
+from ..joint import (
+    reduction_dim_st,
+)
+from ._common import (
+    OpSample,
+    OpSpec,
+)
+
+
+def _reduction_sample_st(
+    method_name: str,
+    allow_keepdim: bool = True,
+) -> st.SearchStrategy[OpSample]:
+    """Sample for a tensor-method reduction taking dim and optional keepdim.
+
+    Drawn jointly: rank in [1, 4], shape, dim in [0, rank), keepdim bool.
+    """
+
+    @st.composite
+    def _draw(draw) -> OpSample:
+        rank = draw(st.integers(min_value=1, max_value=4))
+        shape = tuple(
+            draw(
+                st.lists(
+                    st.integers(min_value=1, max_value=6),
+                    min_size=rank,
+                    max_size=rank,
+                )
+            )
+        )
+        dim = draw(reduction_dim_st(rank))
+        kwargs: T.Dict[str, T.Any] = {"dim": dim}
+        if allow_keepdim:
+            kwargs["keepdim"] = draw(st.booleans())
+        x = draw(
+            tensor_st(
+                shape,
+                torch.float32,
+                finite=True,
+                domain=Interval(-1e2, 1e2),
+            )
+        )
+        return OpSample(
+            inputs=(x,),
+            kwargs={},
+            module=TensorFnPrimitive(method_name, kwargs=kwargs),
+        )
+
+    return _draw()
+
+
+def _sum_full_or_multi_dim_sample_st() -> st.SearchStrategy[OpSample]:
+    """Sum with the full dim surface per torch.sum doc.
+
+    PyTorch's ``torch.sum`` accepts ``dim`` as ``None`` (reduce all),
+    a single int, or a tuple/list of ints (multi-axis reduction). The t2n
+    reducer at ``torch_to_nnef/op/aten/reducer.py:46-53`` handles all
+    three. The original sum-dim spec only swept the single-int case; this
+    one adds the multi-dim and full-reduction surface.
+    """
+
+    @st.composite
+    def _draw(draw) -> OpSample:
+        rank = draw(st.integers(min_value=1, max_value=4))
+        shape = tuple(
+            draw(
+                st.lists(
+                    st.integers(min_value=1, max_value=6),
+                    min_size=rank,
+                    max_size=rank,
+                )
+            )
+        )
+        # Sweep the three dim modes the doc describes.
+        mode = draw(st.sampled_from(["all", "single", "multi"]))
+        kwargs: T.Dict[str, T.Any] = {}
+        if mode == "all":
+            # Two equivalent ways to reduce everything: dim=None or no dim.
+            if draw(st.booleans()):
+                kwargs["dim"] = None
+            # else: don't pass dim at all.
+        elif mode == "single":
+            kwargs["dim"] = draw(st.integers(min_value=0, max_value=rank - 1))
+            kwargs["keepdim"] = draw(st.booleans())
+        else:  # multi
+            n = draw(st.integers(min_value=1, max_value=rank))
+            dims = draw(
+                st.lists(
+                    st.integers(min_value=0, max_value=rank - 1),
+                    min_size=n,
+                    max_size=n,
+                    unique=True,
+                )
+            )
+            kwargs["dim"] = tuple(sorted(dims))
+            kwargs["keepdim"] = draw(st.booleans())
+        x = draw(
+            tensor_st(
+                shape,
+                torch.float32,
+                finite=True,
+                domain=Interval(-1e2, 1e2),
+            )
+        )
+        return OpSample(
+            inputs=(x,),
+            kwargs={},
+            module=TensorFnPrimitive("sum", kwargs=kwargs),
+        )
+
+    return _draw()
+
+
+def _bool_reduction_sample_st(method_name: str) -> st.SearchStrategy[OpSample]:
+    """Any / all reduction over a single dim, bool input + output."""
+
+    @st.composite
+    def _draw(draw) -> OpSample:
+        rank = draw(st.integers(min_value=1, max_value=4))
+        shape = tuple(
+            draw(
+                st.lists(
+                    st.integers(min_value=1, max_value=6),
+                    min_size=rank,
+                    max_size=rank,
+                )
+            )
+        )
+        dim = draw(reduction_dim_st(rank))
+        keepdim = draw(st.booleans())
+        x = draw(tensor_st(shape, torch.bool))
+        return OpSample(
+            inputs=(x,),
+            kwargs={},
+            module=TensorFnPrimitive(
+                method_name, kwargs={"dim": dim, "keepdim": keepdim}
+            ),
+        )
+
+    return _draw()
+
+
+def _prod_dim_sample_st() -> st.SearchStrategy[OpSample]:
+    """Prod reduction. Inputs near 1.0 to keep results bounded."""
+
+    @st.composite
+    def _draw(draw) -> OpSample:
+        rank = draw(st.integers(min_value=1, max_value=4))
+        # Keep dim sizes small (<=4) -- product of many values quickly
+        # under/overflows even at f32.
+        shape = tuple(
+            draw(
+                st.lists(
+                    st.integers(min_value=1, max_value=4),
+                    min_size=rank,
+                    max_size=rank,
+                )
+            )
+        )
+        dim = draw(reduction_dim_st(rank))
+        keepdim = draw(st.booleans())
+        # Values near 1 (in the (0.5, 2.0) range) keep cumulative product
+        # in a numerically stable interval.
+        x = draw(
+            tensor_st(
+                shape,
+                torch.float32,
+                finite=True,
+                domain=Interval(0.5, 2.0),
+            )
+        )
+        return OpSample(
+            inputs=(x,),
+            kwargs={},
+            module=TensorFnPrimitive(
+                "prod", kwargs={"dim": dim, "keepdim": keepdim}
+            ),
+        )
+
+    return _draw()
+
+
+def _var_dim_sample_st() -> st.SearchStrategy[OpSample]:
+    """Var reduction over a single dim, biased estimator only.
+
+    Two t2n limitations narrow this spec for v1:
+
+    1. The var emitter at ``torch_to_nnef/op/aten/math.py:717`` raises
+       NotImplementedError when ``correction != 0`` (PyTorch defaults to
+       1 -- unbiased estimator); we sweep correction=0 only.
+    2. The same emitter does not honor ``keepdim`` -- it always emits a
+       squeezed-axes ``var`` and never reshapes back. ``keepdim=True``
+       would surface as a shape mismatch (ref ``(..., 1, ...)`` vs tract
+       ``(...)``); we sweep keepdim=False only.
+
+    Both are tracked t2n improvements that this spec will widen against
+    once they land.
+    """
+
+    @st.composite
+    def _draw(draw) -> OpSample:
+        rank = draw(st.integers(min_value=1, max_value=4))
+        shape = tuple(
+            draw(
+                st.lists(
+                    st.integers(min_value=2, max_value=6),
+                    min_size=rank,
+                    max_size=rank,
+                )
+            )
+        )
+        dim = draw(reduction_dim_st(rank))
+        x = draw(
+            tensor_st(
+                shape,
+                torch.float32,
+                finite=True,
+                domain=Interval(-1e2, 1e2),
+            )
+        )
+        return OpSample(
+            inputs=(x,),
+            kwargs={},
+            module=TensorFnPrimitive(
+                "var",
+                kwargs={
+                    "dim": dim,
+                    "keepdim": False,
+                    "correction": 0,
+                },
+            ),
+        )
+
+    return _draw()
+
+
+def _reduction_specs() -> T.List[OpSpec]:
+    return [
+        OpSpec(
+            name="sum-dim",
+            sample_st=_reduction_sample_st("sum"),
+            tolerance=TractCheckTolerance.APPROXIMATE,
+            dtypes_hint=(torch.float32,),
+        ),
+        OpSpec(
+            # Multi-dim and full-tensor sums accumulate float error across
+            # many elements; the per-dim form (sum-dim) only sums one axis
+            # so APPROXIMATE works there. Multi-dim sums need CLOSE.
+            name="sum-dim-broad",
+            sample_st=_sum_full_or_multi_dim_sample_st(),
+            tolerance=TractCheckTolerance.CLOSE,
+            dtypes_hint=(torch.float32,),
+        ),
+        OpSpec(
+            name="mean-dim",
+            sample_st=_reduction_sample_st("mean"),
+            tolerance=TractCheckTolerance.APPROXIMATE,
+            dtypes_hint=(torch.float32,),
+        ),
+        OpSpec(
+            name="max-dim",
+            sample_st=_reduction_sample_st("max"),
+            tolerance=TractCheckTolerance.EXACT,
+            dtypes_hint=(torch.float32,),
+        ),
+        OpSpec(
+            name="min-dim",
+            sample_st=_reduction_sample_st("min"),
+            tolerance=TractCheckTolerance.EXACT,
+            dtypes_hint=(torch.float32,),
+        ),
+        # Argmax / argmin return int64 indices -- the comparator's exact
+        # int path catches any divergence. Pure index ops, no tolerance
+        # needed.
+        OpSpec(
+            name="argmax-dim",
+            sample_st=_reduction_sample_st("argmax"),
+            tolerance=TractCheckTolerance.EXACT,
+            dtypes_hint=(torch.float32,),
+        ),
+        OpSpec(
+            name="argmin-dim",
+            sample_st=_reduction_sample_st("argmin"),
+            tolerance=TractCheckTolerance.EXACT,
+            dtypes_hint=(torch.float32,),
+        ),
+        # any / all are bool reductions (input bool, output bool).
+        # tract 0.22.1 (latest in TractNNEF.OFFICIAL_SUPPORTED_VERSIONS)
+        # does NOT define ``any_reduce`` / ``all_reduce`` operators -- they
+        # were added in tract > 0.22.1. The curated test at
+        # ``tests/test_primitive.py:1180-1188`` skips these via
+        # ``cond_tract_gt_0_22_0``. Xfail until the supported version set
+        # bumps past 0.22.1.
+        OpSpec(
+            name="any-dim-xfail",
+            sample_st=_bool_reduction_sample_st("any"),
+            tolerance=TractCheckTolerance.EXACT,
+            dtypes_hint=(torch.bool,),
+            xfail_reason=(
+                "tract 0.22.1 lacks any_reduce; introduced in tract > "
+                "0.22.1. Bumping TractNNEF.OFFICIAL_SUPPORTED_VERSIONS "
+                "will flip this back to a normal pass."
+            ),
+        ),
+        OpSpec(
+            name="all-dim-xfail",
+            sample_st=_bool_reduction_sample_st("all"),
+            tolerance=TractCheckTolerance.EXACT,
+            dtypes_hint=(torch.bool,),
+            xfail_reason=(
+                "tract 0.22.1 lacks all_reduce; introduced in tract > "
+                "0.22.1. Bumping TractNNEF.OFFICIAL_SUPPORTED_VERSIONS "
+                "will flip this back to a normal pass."
+            ),
+        ),
+        # prod is multiplicative reduction; numerical drift accumulates
+        # for long axes and small values, so CLOSE rather than APPROXIMATE.
+        OpSpec(
+            name="prod-dim",
+            sample_st=_prod_dim_sample_st(),
+            tolerance=TractCheckTolerance.CLOSE,
+            dtypes_hint=(torch.float32,),
+        ),
+        # var has unbiased/biased variants via the ``correction`` kwarg.
+        # We sweep both.
+        OpSpec(
+            name="var-dim",
+            sample_st=_var_dim_sample_st(),
+            tolerance=TractCheckTolerance.CLOSE,
+            dtypes_hint=(torch.float32,),
+        ),
+    ]
+
+
+# -----------------------------------------------------------------------------
+# Shape op specs (5)
+# -----------------------------------------------------------------------------
+
+
+def _reduction_dtype_kwarg_sample_st(
+    method_name: str,
+) -> st.SearchStrategy[OpSample]:
+    """Reduction with the ``dtype`` cast-then-reduce kwarg.
+
+    PyTorch's ``sum/mean/prod(dim, *, dtype=)`` casts the input to
+    ``dtype`` BEFORE reducing -- useful for f16 -> f32 accumulation.
+    """
+
+    @st.composite
+    def _draw(draw) -> OpSample:
+        rank = draw(st.integers(min_value=1, max_value=4))
+        shape = tuple(
+            draw(
+                st.lists(
+                    st.integers(min_value=1, max_value=5),
+                    min_size=rank,
+                    max_size=rank,
+                )
+            )
+        )
+        dim = draw(reduction_dim_st(rank))
+        keepdim = draw(st.booleans())
+        # Sweep target dtype: float64 (upcast) and float32 (no-op).
+        target_dtype = draw(st.sampled_from([torch.float32, torch.float64]))
+        x = draw(
+            tensor_st(
+                shape,
+                torch.float32,
+                finite=True,
+                domain=Interval(-1e2, 1e2),
+            )
+        )
+        return OpSample(
+            inputs=(x,),
+            kwargs={},
+            module=TensorFnPrimitive(
+                method_name,
+                kwargs={
+                    "dim": dim,
+                    "keepdim": keepdim,
+                    "dtype": target_dtype,
+                },
+            ),
+        )
+
+    return _draw()
+
+
+def _depth_reduction_dtype_specs() -> T.List[OpSpec]:
+    APPROX = TractCheckTolerance.APPROXIMATE
+    CLOSE = TractCheckTolerance.CLOSE
+    return [
+        OpSpec(
+            name="sum-dtype",
+            sample_st=_reduction_dtype_kwarg_sample_st("sum"),
+            tolerance=APPROX,
+            dtypes_hint=(torch.float32,),
+        ),
+        OpSpec(
+            name="mean-dtype",
+            sample_st=_reduction_dtype_kwarg_sample_st("mean"),
+            tolerance=APPROX,
+            dtypes_hint=(torch.float32,),
+        ),
+        OpSpec(
+            name="prod-dtype",
+            sample_st=_reduction_dtype_kwarg_sample_st("prod"),
+            tolerance=CLOSE,
+            dtypes_hint=(torch.float32,),
+        ),
+    ]
+
+
+# -----------------------------------------------------------------------------
+# Registry assembly
+# -----------------------------------------------------------------------------
